requirements.txt
django==3.2.7
djangorestframework=3.12.4
